+++ src/components/Joyride/Joyride.helpers.js
